day 18; updated extension methods

+++ AdventOfCode2016/Extensions/StringExtensions.cs
@@ -9,8 +9,8 @@ namespace AdventOfCode2016.Extensions
     {
         public static bool IsNumeric(this string s)
         {
-            int a;
-            return int.TryParse(s, out a);
+            long a;
+            return long.TryParse(s, out a);
         }
 
         public static int ToInt(this string s)
